Allow for dropna, targeted orgs & refactor
Refactor includes dropping examples with mock client, removing example
data, a Jupyter notebook, adding .dockerignore, and renaming the module
to avoid issues with package names containing a hyphen ('-'), which
Python does not allow in import statements.

Running the average algorithm is a common way of testing the
infrastructure, but we don't always want to run it against every node in
the collaboration. Allowing for targeting specific organizations is
desirable.

Jupyter notebooks are hard to maintain in Git repositories and make it
difficult to understand what's being committed into the tree.

The mock client lacks support for some features (e.g., environment
variables in node config). Standing up local nodes and a server would be
a better option, although this is not as easy as it be should at the
moment.